Increase background spread threshold back to 0.2 and added checkbox for removing datasets from SEC tab.  Datasets that are not inUse should be excluded from SEC builder.  This is acheived by removing unchecked datasets from the collection.

+++ src/version4/SEC/SECFile.java
@@ -250,7 +250,7 @@ public class SECFile {
                 last.add(Double.parseDouble(fIndex), Double.parseDouble(signals[i]));
                 signalSeries.add(last.getDataItem(0));
 
-                if (Integer.parseInt(bckgrnd[i]) > 0){
+                if (Double.parseDouble(bckgrnd[i]) > 0){
                     selectedBuffers.add(last.getDataItem(0));
                 }
             }
